+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New features
 
+* Adds new `DataCombined` object that combines observed data (`DataSet`) and simulated data (`SimulationResults`) into a single object. Especially important is the `$toDataFrame()` method of this object that returns a dataframe containing combined data from observed and simulated data, which can be further used for visualizations or other analysis.
+
 * Adds `simulationResultsToDataFrame()` function to convert `SimulationResults` objects into a dataframe.
 
 ## Breaking changes